minor fixes

+++ fms-prototype-3/booking-service/src/main/java/bookings/web/rest/BookingResource.java
@@ -3,6 +3,7 @@ package bookings.web.rest;
 import bookings.domain.Booking;
 import bookings.domain.enumeration.ETopicType;
 import bookings.repository.BookingRepository;
+import bookings.security.SecurityUtils;
 import bookings.service.BookingKafkaProducer;
 import bookings.web.rest.errors.BadRequestAlertException;
 
@@ -18,6 +19,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -98,7 +100,17 @@ public class BookingResource {
     @GetMapping("/bookings")
     public List<Booking> getAllBookings() {
         log.debug("REST request to get all Bookings");
-        return bookingRepository.findAll();
+        List<Booking> resultingBookings = new ArrayList<>();
+        // Return 404 if the entity is not owned by the connected user
+        Optional<String> userLogin = SecurityUtils.getCurrentUserLogin();
+
+        for(Booking b: bookingRepository.findAll()) {
+            if(userLogin.isPresent() && userLogin.get().equals(b.getPassengerId())) {
+                resultingBookings.add(b);
+            }
+        }
+
+        return resultingBookings;
     }
 
     /**
@@ -111,7 +123,16 @@ public class BookingResource {
     public ResponseEntity<Booking> getBooking(@PathVariable Long id) {
         log.debug("REST request to get Booking : {}", id);
         Optional<Booking> booking = bookingRepository.findById(id);
-        return ResponseUtil.wrapOrNotFound(booking);
+
+        // Return 404 if the entity is not owned by the connected user
+        Optional<String> userLogin = SecurityUtils.getCurrentUserLogin();
+        if(booking.isPresent() &&
+            userLogin.isPresent() &&
+            userLogin.get().equals(booking.get().getPassengerId())) {
+            return ResponseUtil.wrapOrNotFound(booking);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     /**
@@ -124,9 +145,20 @@ public class BookingResource {
     public ResponseEntity<Void> deleteBooking(@PathVariable Long id) {
         log.debug("REST request to delete Booking : {}", id);
 
-        bookingRepository.deleteById(id);
-        bookingKafkaProducer.sendBookingEvent(bookingRepository.findById(id).get(), ETopicType.CANCELLED);
+        Optional<Booking> booking = bookingRepository.findById(id);
 
+        // Return 404 if the entity is not owned by the connected user
+        Optional<String> userLogin = SecurityUtils.getCurrentUserLogin();
+        if(booking.isPresent() &&
+            userLogin.isPresent() &&
+            userLogin.get().equals(booking.get().getPassengerId())) {
+            bookingRepository.deleteById(id);
+            if(bookingRepository.findById(id).isPresent()) {
+                bookingKafkaProducer.sendBookingEvent(bookingRepository.findById(id).get(), ETopicType.CANCELLED);
+            }
+        } else {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.noContent().headers(HeaderUtil.createEntityDeletionAlert(applicationName, false, ENTITY_NAME, id.toString())).build();
     }
 }
